Added frameType.ts : resturct constants folder
